Fix anti join w/ filter that is true or failing for non-matching rows (#3571)
Summary:
Anti join returned incorrect result or failed when extra filter passed or failed on rows without a match.

The fix is to not evaluate the filter on rows without a match and always include these in the join results.

A similar change needs to be applies in a follow-up PR to the left join. The
filter should not be evaluated on rows without a match to avoid failing the
query unnecessarily.

See #3555 for addition context.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/3571

Reviewed By: xiaoxmeng

Differential Revision: D42194249

Pulled By: mbasmanova

fbshipit-source-id: eccb52de2ebb72e2c355d53590ee138efb16cdec